Start to make e2e test more manageable
