+++ e2e/nx-firebase-e2e/test-utils/test-helpers.ts
@@ -1,0 +1,11 @@
+export function expectStrings(input: string, contains: string[]) {
+  contains.forEach((item) => {
+    expect(input).toContain(item)
+  })
+}
+
+export function expectNoStrings(input: string, contains: string[]) {
+  contains.forEach((item) => {
+    expect(input).not.toContain(item)
+  })
+}
